union pars scripts

+++ parse.py
@@ -3,7 +3,7 @@ import json
 from secrets import phone_number, password, owner_id, password_db
 from datetime import datetime
 import psycopg2
-
+# FIXME ЭТОТ КОД НЕ ТЕСТИРОВАЛСЯ, КОММИТ ПОКАЗЫВАЕТ ТОЛЬКО ОБЩИЙ ПРИНЦИП И МОЖЕТ НЕ РАБОТАТЬ
 # создание сессии
 vk_session = vk_api.VkApi(phone_number, password, app_id=2685278)
 vk_session.auth(token_only=True)
@@ -19,13 +19,35 @@ now = str(datetime.now().strftime('%Y-%m-%d %H:%M:%S'))
 reply = vk.wall.get(owner_id=owner_id, count=0)
 total_count = reply['count']
 
+newDB = false
+old_total_count = 0
+data = None
+# >>> Мы ПЫТАЕМСЯ прочитать файл с данными
+try:
+  with open('posts.json', 'r', encoding='utf-8') as outfile:
+    data = json.load(outfile)
+    old_total_count = data['total_count']
+    # FIXME: Костыль (возможно)
+    newDB = true
+except:
+    print("Can't find previous parsing data")
+finally:
+    # >>> По окончанию блока подставляем данные о тотале
+    # считаем, сколько постов добавилось, и задаем max_count
+    count = total_count - old_total_count
+    max_count = count // 25 + 1
+
 # парсим посты и переворачиваем список с ними, потому что нам нужны данные от старых к новым, а парсятся
 # от новых к старым
+# также обрезаем список до нужного нам кол-ва постов, так как get_all получает кол-во постов, кратное 25
 reply = tools.get_all(method='wall.get', max_count=1, values={'owner_id': owner_id}, limit=1)
+items = reply['items'][:count:]
 items = reply['items'][::-1]
 
-# создаем список, куда запихнем посты
-posts = []
+# неприятно костыльно создаем два списка, в posts сохраняем все прошлые посты из data, потом туда добавим новые
+# posts_db оставляем пустым, потом туда добавим также новые (но там будут только они)
+posts = data['items']
+posts_db = []
 
 # функция, чтобы обернуть комменты просто в список текстов
 def get_comments(items):
@@ -54,8 +76,9 @@ for item in items:
         post['comments'] = {'count': len(comments), 'items': comments}
         post['attachments'] = len(item['attachments'])
 
-        # добавляем пост в posts
+        # добавляем пост в posts и posts_db
         posts.append(post)
+        posts_db.append(post)
 
 # собираем result, который пойдет в json
 result = {'total_count': total_count, 'date_updated': now, 'items': posts}
@@ -68,17 +91,27 @@ with open('posts.json', 'w', encoding='utf-8') as outfile:
 conn = psycopg2.connect(database="parsed_posts", user="postgres", password=password_db, host="localhost", port="5432")
 
 cur = conn.cursor()
+res = None
+if newDB:
+    # очищаем таблицы и сбрасываем PK
+    cur.execute('TRUNCATE TABLE comments RESTART IDENTITY CASCADE')
+    cur.execute('TRUNCATE TABLE posts RESTART IDENTITY CASCADE')
+else:
+    # вытаскиваем номер последнего на данный момент времени поста в бд,
+    # чтобы потом дать правильные post_id комментам в их таблице
+    cur.execute('SELECT post_id FROM posts ORDER BY post_id DESC LIMIT 1')
+    res = cur.fetchone()
 
-# очищаем таблицы и сбрасываем PK
-cur.execute('TRUNCATE TABLE comments RESTART IDENTITY CASCADE')
-cur.execute('TRUNCATE TABLE posts RESTART IDENTITY CASCADE')
-
-# тут без проверки говна, просто i
-i = 1
+# проверка говна, которая будет нужна, только если парсер будет один общий
+if res is not None:
+    i = res[0]
+else:
+    i = 1
 
 # просто записываем новые посты в бд
 # юзаем posts
-for post in posts:
+# FIXME предположу что тут нужен posts_db
+for post in posts_db:
     cur.execute('INSERT INTO posts (views, likes, reposts, attachments, text, date) VALUES (%s, %s, %s, %s, %s, %s)',
                 (post['views'], post['likes'], post['reposts'], post['attachments'], post['text'], post['date']))
 
@@ -86,7 +119,7 @@ conn.commit()
 
 # записываем комменты новых постов в бд
 # опять же юзаем posts
-for post in posts:
+for post in posts_db:
     # вытаскиваем комменты из поста
     comments = post['comments']['items']
